# api/v1/openapi.yml
openapi: 3.0.3
info:
  title: Sketch API on GO
  description: Тестовый API сервис написанный на Go.
  version: "1.0.0"
servers:
  - url: /api/v1
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    RegisterUserRequest:
      type: object
      required: [username, email, password, confirmation]
      properties:
        username:
          type: string
          minLength: 1
        email:
          type: string
          format: email
          maxLength: 254
        password:
          type: string
          format: password
          minLength: 8
          maxLength: 128
        confirmation:
          type: string
          format: password
          minLength: 8
          maxLength: 128
...
    UserResponse:
      type: object
      required: [id, username]
      properties:
        id:
          type: string
          format: uuid
        username:
          type: string
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              enum:
                - INVALID_REQUEST
                - UNAUTHORIZED
                - NOT_FOUND
                - FORBIDDEN
                - SCHEDULE_EXISTS
                - INTERNAL_ERROR
            message:
              type: string
      example:
        error:
          code: INVALID_REQUEST
          message: invalid request
    InternalErrorResponse:
      type: object
      required: [message]
      properties:
        message:
          type: string
paths:
...
  /user/register:
    post:
      tags: [User]
      summary: "Регистрация пользователя."
      description: "..."
      operationId: registerUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterUserRequest'

      responses:
        '201':
          description: Пользователь создан
          content:
            application/json:
              schema:
                 $ref: '#/components/schemas/UserResponse'

        '400':
          description: Неверный запрос
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }

        '409':
          description: Пользователь уже существует
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

        '500':
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema: { $ref: '#/components/schemas/InternalErrorResponse' }
...